API: Less aggressive API shift, create TyphonSuite
