+++ typhon/__init__.py
@@ -1,6 +1,7 @@
 __all__ = ['TyphonDisplay', 'DeviceDisplay', 'use_stylesheet',
-           'register_signal']
+           'register_signal', 'TyphonSuite']
 from .display import TyphonDisplay, DeviceDisplay
+from .suite import TyphonSuite
 from .utils import use_stylesheet
 from .plugins import register_signal
 from ._version import get_versions
