+++ typhon/display.py
@@ -1,32 +1,24 @@
-############
-# Standard #
-############
 import logging
 import os.path
+from warnings import warn
 
-############
-# External #
-############
 from ophyd import Device
+from pydm.widgets.drawing import PyDMDrawingImage
 from qtpy import uic
-from qtpy.QtCore import Slot, Qt, QModelIndex
+from qtpy.QtCore import Qt
 
-###########
-# Package #
-###########
-from .device import TyphonPanel
-from .utils import ui_dir, clean_name, TyphonBase
-from .widgets import TyphonSidebarItem
-from .tools import TyphonTimePlot, TyphonLogDisplay
+from .func import FunctionPanel
+from .signal import SignalPanel
+from .utils import ui_dir, clean_attr, clean_name, TyphonBase
 
 logger = logging.getLogger(__name__)
 
 
 class TyphonDisplay(TyphonBase):
     """
-    Generalized Typhon display
+    Main Panel display for a signal Ophyd Device
 
-    This widget lays out all of the architecture for a general Typhon display.
+    This widget lays out all of the architecture for a single Ophyd display.
     The structure matches an ophyd Device, but for this specific instantation,
     one is not required to be given. There are four main panels available;
     :attr:`.read_panel`, :attr:`.config_panel`, :attr:`.method_panel`. These
@@ -36,247 +28,145 @@ class TyphonDisplay(TyphonBase):
     default, widgets are hidden until contents are added. For instance, if you
     do not add any methods to the main panel it will not be visible.
 
-    This device is the bare bones implementation in the event that someone
-    might want to collect a random group of signals and devices together to
-    create a screen. A more automated display generator is available  in the
-    :class:`.DeviceDisplay`
+    This contains the widgets for all of the root devices signals, any methods
+    you would like to display, and an optional image. As with ``typhon``
+    convention, the base initialization sets up the widgets and the
+    ``.from_device`` class method will automatically populate them.
 
     Parameters
     ----------
-    name : str
-        Title displayed on the widget
+    name: str, optional
+        Name to displayed at the top of the panel
 
-    image :str, optional
-        Path to an image file to include in the display.
+    image: str, optional
+        Path to image file to displayed at the header
 
-    parent : QWidget, optional
+    parent: QWidget, optional
     """
-
-    def __init__(self, parent=None):
-        # Instantiate Widget
+    def __init__(self, name=None, image=None, parent=None):
         super().__init__(parent=parent)
         # Instantiate UI
-        self.ui = uic.loadUi(os.path.join(ui_dir, 'base.ui'), self)
-        self.device_panel = TyphonPanel()
-        self.widget_layout.insertWidget(1, self.device_panel)
-        self.subdisplays = dict()
-        # Connect signals to slots
-        self.ui.hide_button.clicked.connect(self.hide_subdisplays)
-        self.ui.tool_list.clicked.connect(self.show_subdisplay)
-        self.ui.tool_list.clicked.connect(
-                self.ui.component_list.clearSelection)
-        self.ui.component_list.clicked.connect(self.show_subdisplay)
-        self.ui.component_list.clicked.connect(
-                self.ui.tool_list.clearSelection)
-        # Hide widgets until objects are added to them
-        self.ui.subwindow.hide()
-        self.ui.tool_sidebar.hide()
-        self.ui.component_sidebar.hide()
-
-    def add_subdisplay(self, name, display, list_widget):
-        """
-        Add a widget to one of the button layouts
-
-        This add a display for a subcomponent and a QPushButton that
-        will bring the display to the foreground. Users can either specify
-        their button or have one generate for them. Either way the button is
-        connected to the `pyqSlot` :meth:`.show_subdevice`
-
-        Parameters
-        ----------
-        name : str
-            Name to place on QPushButton
-
-        display : QWidget
-            QWidget to associate with button
-
-        button : QWidget, optional
-            QWidget with the PyQtSignal ``clicked``. If None, is given a
-            QPushButton is created
-        """
-        # Create QListViewItem to store the display information
-        list_item = TyphonSidebarItem(name)
-        list_item.setData(Qt.UserRole, display)
-        list_widget.addItem(list_item)
-        # Add our display to the component widget
-        self.ui.subdisplay.addWidget(display)
-        self.subdisplays[name] = list_item
-        # Hide the parent widget if hidden
-        sidebar = list_widget.parent()
-        if sidebar.isHidden():
-            sidebar.show()
-
-    def add_subdevice(self, device, name=None, **kwargs):
-        """
-        Add a subdevice to the `component_widget` stack
-
-        Parameters
-        ----------
-        device : ophyd.Device
-
-        kwargs:
-            Passed to :class:`.TyphonPanel` constructor
-        """
-        logger.debug("Creating subdisplay for %s", device.name)
-        # Remove parent from name for title
-        if not name:
-            name = clean_name(device)
-        dd = TyphonPanel.from_device(device, name=name, **kwargs)
-        self.add_subdisplay(name, dd, self.component_list)
-
-    def add_tool(self, name, tool):
-        """
-        Add a widget to the toolbar
-
-        Parameters
-        ----------
-        name :str
-            Name of tool to be displayed in sidebar
-
-        tool: QWidget
-            Widget to be added to ``.ui.subdisplay``
-        """
-        self.add_subdisplay(name, tool, self.ui.tool_list)
-
-    def _item_from_sidebar(self, name):
-        """
-        Get a child display based on the name given
-
-        Parameters
-        ----------
-        name: str
-            Name of display
-        """
-        # Gather QListWidgetItem
-        try:
-            list_item = self.subdisplays[name]
-        except KeyError as exc:
-            raise ValueError("Display {} has not been added to the "
-                             "DeviceDisplay yet".format(name)) from exc
-        return list_item
-
-    def get_subdisplay(self, display):
-        """
-        Get a subdisplay by name
-
-        Parameters
-        ----------
-        display :str or Device
-            Name of subdisplay. This will be the text shown on the sidebar. For
-            component devices screens you can pass in the component device
-            itself
-
-        Returns
-        -------
-        widget : QWidget
-            Widget that is a member of the :attr:`.ui.subdisplay`
-
-        Example
-        -------
-        .. code:: python
-
-            my_display.get_subdisplay(my_device.x)
-            my_displyay.get_subsdisplay('My Tool')
-        """
-        # Get the cleaned Device name if passed a Device
-        if isinstance(display, Device):
-            display = clean_name(display)
-        return self._item_from_sidebar(display).data(Qt.UserRole)
-
-    @Slot(str)
-    @Slot(QModelIndex)
-    def show_subdisplay(self, item):
-        """
-        Show subdevice display of the QStackedWidget
-
-        Parameters
-        ----------
-        name : str, Device or QModelIndex
-        """
-        # Grab the relevant display
-        if isinstance(item, QModelIndex):
-            display = item.data(Qt.UserRole)
-        else:
-            display = self.get_subdisplay(item)
-        # Show our subdisplay if previously hidden
-        if self.ui.subwindow.isHidden():
-            self.ui.subwindow.show()
-        # Set the current display
-        self.ui.subdisplay.setCurrentWidget(display)
-        self.ui.subdisplay.setFixedWidth(display.sizeHint().width())
-
-    @Slot()
-    def hide_subdisplays(self):
-        """
-        Hide the component widget and set all buttons unchecked
-        """
-        # Hide the main display
-        self.ui.subwindow.hide()
-        self.ui.component_list.clearSelection()
-        self.ui.tool_list.clearSelection()
+        self.ui = uic.loadUi(os.path.join(ui_dir, 'device.ui'), self)
+        # Set Label Names
+        if name:
+            self.title = name
+        # Create child panels
+        self.method_panel = FunctionPanel()
+        self.read_panel = SignalPanel()
+        self.config_panel = SignalPanel()
+        self.misc_panel = SignalPanel()
+        # Add to layout
+        self.read_widget.setLayout(self.read_panel)
+        self.config_widget.setLayout(self.config_panel)
+        self.misc_widget.setLayout(self.misc_panel)
+        self.main_layout.insertWidget(3, self.method_panel,
+                                      0, Qt.AlignHCenter)
+        self.method_panel.hide()
+        # Create PyDMDrawingImage
+        self.image_widget = None
+        if image:
+            self.add_image(image)
 
     @property
-    def tools(self):
-        """Tools loaded into the DeviceDisplay"""
-        return [self.tool_list.item(i).data(Qt.UserRole)
-                for i in range(self.tool_list.count())]
+    def title(self):
+        """Title at the top of panel"""
+        return self.name_label.text()
 
-    def add_device(self, device, children=True, methods=None, image=None):
+    @title.setter
+    def title(self, text):
+        self.name_label.setText(text)
+
+    @property
+    def methods(self):
         """
-        Add a device to the :attr:`.device_panel` and tools
+        Methods contained within :attr:`.method_panel`
+        """
+        return self.method_panel.methods
+
+    def add_image(self, path):
+        """
+        Set the image of the PyDMDrawingImage
+
+        Setting this twice will overwrite the first image given.
+
+        Parameters
+        ----------
+        path : str
+            Absolute or relative path to image
+        """
+        # Set existing image file
+        logger.debug("Adding an image file %s ...", path)
+        if self.image_widget:
+            self.image_widget.filename = path
+        else:
+            logger.debug("Creating a new PyDMDrawingImage")
+            self.image_widget = PyDMDrawingImage(filename=path,
+                                                 parent=self)
+            self.image_widget.setMaximumSize(350, 350)
+            self.ui.main_layout.insertWidget(2, self.image_widget,
+                                             0, Qt.AlignCenter)
+
+    def add_device(self, device, methods=None):
+        """
+        Add a Device and signals to the TyphonDisplay
 
         Parameters
         ----------
         device: ophyd.Device
 
         methods: list, optional
-            Methods to add to the device
+            List of methods to add to the :attr:`.method_panel`
         """
         super().add_device(device)
-        # Add the device to the main panel
-        self.device_panel.add_device(device, methods=methods)
-        self.device_panel.add_image(image)
-        # Add a device to all the tool displays
-        for tool in self.tools:
-            try:
-                tool.add_device(device)
-            except Exception as exc:
-                logger.exception("Unable to add %s to tool %s",
-                                 device.name, type(tool))
-        # Handle child devices
-        if children:
-            for dev_name in device._sub_devices:
-                self.add_subdevice(getattr(device, dev_name))
+        # Create read and configuration panels
+        for attr in device.read_attrs:
+            signal = getattr(device, attr)
+            if not isinstance(signal, Device):
+                self.read_panel.add_signal(signal, clean_attr(attr))
+        for attr in device.configuration_attrs:
+            signal = getattr(device, attr)
+            if not isinstance(signal, Device):
+                self.config_panel.add_signal(signal, clean_attr(attr))
+        # Catch the rest of the signals add to misc panel below misc_button
+        for attr in device.component_names:
+            if attr not in (device.read_attrs
+                            + device.configuration_attrs):
+                signal = getattr(device, attr)
+                if not isinstance(signal, Device):
+                    self.misc_panel.add_signal(signal, clean_attr(attr))
+        # Add our methods to the panel
+        methods = methods or list()
+        for method in methods:
+                self.method_panel.add_method(method)
+
+    def add_tab(self, name, widget):
+        warn("This method will be deprecated in a future release",
+             category=DeprecationWarning)
+        self.signal_tab.add_tab(widget, name)
 
     @classmethod
-    def from_device(cls, device, parent=None,
-                    tools={'Log': TyphonLogDisplay,
-                           'StripTool': TyphonTimePlot},
-                    **kwargs):
+    def from_device(cls, device, methods=None, name=None, **kwargs):
         """
-        Create a new TyphonDisplay from an ophyd.Device
+        Create a new TyphonDisplay from a Device
+
+        Loads the signals in to the appropriate positions and sets the title to
+        a cleaned version of the device name
 
         Parameters
         ----------
         device: ophyd.Device
 
-        children: bool, optional
-            Choice to include child Device components
-
-        parent: QWidgets
+        methods: list
+            List of methods to add to the :attr:`.method_panel`
 
         kwargs:
-            Passed to :meth:`.add_device`
+            Passed TyphonDisplay constructor
         """
-        display = cls(parent=parent)
-        for name, tool in tools.items():
-            try:
-                display.add_tool(name, tool())
-            except Exception:
-                logger.exception("Unable to load %s", type(tool))
-        display.add_device(device, **kwargs)
-        display.device_panel.title = clean_name(device, strip_parent=False)
-        return display
+        if not name:
+            name = clean_name(device, strip_parent=False)
+        panel = cls(name=name, **kwargs)
+        panel.add_device(device, methods=methods)
+        return panel
 
 
 DeviceDisplay = TyphonDisplay.from_device
